Created a new `defGalaxy.py` because it was easier then fixing the old one.
The original `defGalaxy.py` is now `defGalaxy_old.py`. It will be deleted
one `defGalaxy.py` gets build (currently it is gutless). In comparison
`defGalaxy-new.py` was a test case for the effects of smoothing and box-
adding pixels on galaxy definition. Its results will be incorporated into
`defGalaxy.py`.