Start of major overhaul and reorganization.
Messed with a lot of comments blocks to make it to my new standards.
Worked on `get_SN_SDSS_coord()` and `get_SN_HST_coord()`. These are a
split of the old `get_SN_position()`. `get_SN_SDSS_coord()` is finished,
but `get_SN_HST_coord()` needs some the SDSS-to-HST shift applied.