Adjusted the computation of FPR to work with more possible SN values.
It now works sanely if the SN's pixel value is found once in the array of
galactic values, if its found multiple times, and if it is not found at
all (aka between galactic values). The last case will be the most likely
case used because we are going to be calculating the SN's pixel value from
a 3x3 grid around the SN's location.