Update fpr pipeline to correctly use ACS/WFC
Notes
======

SN19282 got lost from HST. IT was found with (19282, 60  1029.61359247,
1049.13015776, 2.2550640106, 1.9514168501, 1.5560578108) before the change in
the pixel size increased the threshold in flux. Now the guess is (19282,
52  1031.16632261, 985.207062756, 1.9911431074, 1.8429368734, 1.3052073717).

SN20350 might need to be deleted form analysis because it has an object
bleeding  over the host, especially obvious in the SDSS images. If so, add
this to line 355 in `fractionalRank.py`.